VersionInfo is in version.h
This fixes the constant re-compilation caused by having gitinfo.h imported in basics.h
